requirements.txt
Pillow
psycopg
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
pyyaml>=6.0
numpy>=1.24.0
python-dotenv
gunicorn
whitenoise
google-generativeai
